Add ZIP code logic
